+++ app/src/main/java/com/sopt/now/data/model/request/RequestSignupDto.kt
@@ -1,12 +1,16 @@
-package com.sopt.now.data.DTO.request
+package com.sopt.now.data.model.request
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class RequestLoginDto(
+data class RequestSignupDto(
     @SerialName("authenticationId")
     val authenticationId: String,
     @SerialName("password")
     val password: String,
+    @SerialName("nickname")
+    val nickname: String,
+    @SerialName("phone")
+    val phone: String,
 )
