+++ app/src/main/java/com/sopt/now/data/model/response/ResponseUserInfoDto.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.data.DTO.response
+package com.sopt.now.data.model.response
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
