+++ app/src/main/java/com/sopt/now/presentation/Key.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.data
+package com.sopt.now.presentation
 
 object Key {
     const val PW = "pw"
